add jpretty; mega update to moonphase and greetings; some general patches

+++ functions/scs.ps1
@@ -20,7 +20,10 @@ function scs {
         
         # open latest screenshot
         [Alias('o')]
-        [switch] $OpenFile
+        [switch] $OpenFile,
+
+        [Alias('k')]
+        [switch] $Keep
     )
 
     # -- setup
@@ -29,22 +32,21 @@ function scs {
     $dateToday = (Get-Date)
     $date = $dateToday.ToString('yyyy-MM\\dd')
     $folderpath = "$pwsh_screenshotpath\$date"
-
     
     # open folder if no options selected
-    if (-not ($OpenFolder -or $Copy -or $OpenFile)) {
+    if (-not ($OpenFolder -or $Copy -or $OpenFile -or $Keep)) {
         $OpenFolder = $true
     }
 
     # -- get latest screenshot folder
 
     # search for previous dates if folder doesn't exist
-    wr "Checking for today's screenshot folder..." -n
+    wr "checking for today's screenshot folder..." -n
     for ($offset = 1; (!(Test-Path $folderpath)); $offset++) {
         # output for first vs subsequent missing dates
         if ($offset -eq 1) {
             wr ""
-            wr "$folderpath not found! Searching for most recent screenshot folder:" -f yellow
+            wr "$folderpath not found! searching for most recent screenshot folder:" -f yellow
 
         } else {
             wr " not found."
@@ -59,7 +61,7 @@ function scs {
         $date = $dateToday.AddDays(-$offset).ToString('yyyy-MM\\dd')
         $folderpath = "$pwsh_screenshotpath\$date"
 
-        wr "Checking " -n
+        wr "checking " -n
         wr "$date..." -f white -n
     }
     wr " screenshot folder $date exists." -f green
@@ -68,19 +70,19 @@ function scs {
     
     # open latest located screenshot folder
     if ($OpenFolder) {
-        wr "Opening screenshot folder " -n
+        wr "opening screenshot folder " -n
         wr "$date... " -f white -n
         explorer $folderpath
         wr "done!" -f green
     }
 
     # exit function if only opening folder
-    if (-not ($Copy -or $OpenFile)) {
+    if (-not ($Copy -or $OpenFile -or $Keep)) {
         return
     }
     
     # get latest screenshot for copy/open
-    wr "Fetching latest file in screenshot folder " -n
+    wr "fetching latest file in screenshot folder " -n
     wr "$date" -f white -n
     wr "..." -n
 
@@ -95,16 +97,22 @@ function scs {
     
     wr "done!" -f green
 
+    if ($Keep) {
+        ro "moving |@b|$($screenshot.Name) |@|to keep... " -n
+        mv $screenshot.FullName "$pwsh_screenshotpath\keep\"
+        ro "|@s|done!"
+    }
+
     # set clipboard to screenshot
     if ($Copy) {
-        wr "Copying image '$($screenshot.Name)' to clipboard... " -f cyan -n
+        wr "copying image '$($screenshot.Name)' to clipboard... " -f cyan -n
         gi $screenshot.FullName | Set-Clipboard
         wr "done!" -f green
     }
 
     # open screenshot
     if ($OpenFile) {
-        wr "Opening '$($screenshot.Name)'... " -n
+        wr "opening '$($screenshot.Name)'... " -n
         start $screenshot.FullName
         wr "done!" -f green
     }
@@ -113,6 +121,10 @@ function scs {
 function scs-clear {
     wr "removing screenshot folders older than 3 months!" -f yellow
     foreach ($folder in (gci $pwsh_screenshotpath |? LastWriteTime -lt ((Get-Date).AddMonths(-3)))) {
+        # skip keep folder
+        if ($folder.Name -eq 'keep') {
+            continue
+        }
         wr "removing screenshots from $($folder.Name)... " -n
         Remove-Item $folder -Recurse -Force -Confirm:$false -ErrorAction Stop | Out-Null
         wr "done!" -f white
